Cleanup channel's "in" file before trying to join
It might(and it did) happen that channel's "in" wasn't pipe, but file.
This might be because something is "stale", ii got killed and couldn't
cleanup or bug in iibot-ng.

We do this with main "in" pipe(file), I see no reason why we couldn't do
it for channels as well.